Renamed files to be consistent style
